requirements.txt
aiogram~=3.24.0
python-dotenv~=1.2.1
pandas~=2.3.3
requests~=2.32.5
openpyxl==3.1.2
aiohttp==3.9.0